pyproject.toml
[tool.poetry]
name = "remainder-bot"
version = "0.1.0"
description = ""
authors = ["boronoff <romanticboron@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.15.0"
apscheduler = "^3.11.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
